+++ Client/app/src/main/java/com/labs/valtech/vris/business/factories/DataModel/DataModelFactory.kt
@@ -1,0 +1,8 @@
+package com.labs.valtech.vris.business.factories.DataModel
+
+/**
+ * Created by marvin.brouwer on 27-12-2017.
+ */
+sealed class DataModelFactory : IDataModelFactory{
+
+}
